Moving ES-specific parts out of `ViewsBindings`.
